% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createHeatmap}
\alias{createHeatmap}
\title{Create Heatmap}
\usage{
createHeatmap()
}
\value{
This function returns a highcharter object (heatmap).
}
\description{
This function creates a heatmap of the priority matrix available in \link{createPriorityMatrix}.
}
\examples{
createHeatmap()

}
