% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createParetoChart}
\alias{createParetoChart}
\title{Create Pareto Chart}
\usage{
createParetoChart(pOptions = c("Customer 1", "Customer 2", "Customer 3"),
  poImport = c(5, 3, 1))
}
\arguments{
\item{pOptions}{The names of the Pareto options in \link{readParetoOptions}.}

\item{poImport}{The numeric values assigned to each Pareto option in \link{paretoImportance}.}
}
\value{
This function returns a Pareto chart as a highcharter object (column and line charts).
}
\description{
This function creates a Pareto chart of the Pareto options evaluated in \link{paretoImportance}.
}
\examples{
createParetoChart(pOptions,poImport)

}
