% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createPriorityMatrix}
\alias{createPriorityMatrix}
\title{Create Priority Matrix}
\usage{
createPriorityMatrix(vImport = c(3, 3, 5), pImport = c(2, 5, 5),
  priorities = c("Professional certifications", "On the job learning",
  "Mentorship"), values = c("Contribution", "Certainty", "Connection"))
}
\arguments{
\item{vImport}{a numeric list assigning importance to values.
Can be taken from \link{valuesImportance}.}

\item{pImport}{a numeric list assigning importance to priorities
Can be taken from \link{priorityImportance}.}

\item{priorities}{a character array of priorities.
Can be taken from \link{readPriorities}.}

\item{values}{a character array of values
Can be taken from \link{readValues}.}
}
\value{
This function returns a priority matrix as a data frame (values as columns and priorities as rows).  
Values in matrix are the result of cross-multiplication.
}
\description{
This function creates a priority matrix that is a result of \link{priorityImportance} and 
\link{valuesImportance} for numeric values of \link{readValues} and \link{readPriorities}.
}
\examples{
createPriorityMatrix(vImport,pImport,priorities,values)

}
