% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createStrategyCanvas}
\alias{createStrategyCanvas}
\title{Create Strategy Canvas}
\usage{
createStrategyCanvas(sOptions = c("Option 1", "Option 2", "Option 3"),
  soImport1 = c(5, 3, 1), soImport2 = c(1, 3, 5))
}
\arguments{
\item{sOptions}{The names of the strategy options in \link{readStrategyOptions}.}

\item{soImport1}{The numeric values assigned to each strategy option relative to industry standards
in \link{strategyOptionImportance}.}

\item{soImport2}{The numeric values assigned to each strategy option relative to differentiated standards
in \link{strategyOptionImportance}.}
}
\value{
This function returns a strategy canvas as a highcharter object (two line charts).
}
\description{
This function creates a strategy canvas of the strategy options evaluated in \link{strategyOptionImportance}.
}
\examples{
createStrategyCanvas(sOptions,soImport1,soImport2)

}
