% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{optionImportance}
\alias{optionImportance}
\title{Evaluate a Strategic Option Against Priority Matrix}
\usage{
optionImportance(arrayList = c("Professional certifications",
  "On the job learning", "Mentorship"))
}
\arguments{
\item{arrayList}{a list of priorities to evaluate numeric significance for a given strategic option.
Can be taken from \link{readPriorities} output object of \code{priorities}.}
}
\value{
This function returns a data frame to evaluate strategic option alignment with a given set of priorities
and values.
}
\description{
This function determines whether a given strategic option is aligned with priorities as defined by the
\link{createPriorityMatrix} and \link{createHeatmap}.
}
\examples{
optionImportance(priorities)
# Enter the name of the option being considered:
Job from Company A
# On a scale of 1 to 5, how much does Job.from.Company.A fulfill Professional certifications ?
1
# On a scale of 1 to 5, how much does Job.from.Company.A fulfill On the job learning ?
5
# On a scale of 1 to 5, how much does Job.from.Company.A fulfill Mentorship ?
5

}
