% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paretoImportance}
\alias{paretoImportance}
\title{Rank Importance of Pareto Options}
\usage{
paretoImportance(arrayList = c("Customer 1", "Customer 2", "Customer 3"))
}
\arguments{
\item{arrayList}{a list of values to assign numeric importance.  Can be taken from \link{readParetoOptions}.}
}
\value{
This function does not return any values.  Instead, it saves an array numeric object based
on the user's input to the global environment.
}
\description{
This function assigns relative rank and importance to the Pareto options defined in \link{readParetoOptions}.  
These numerical assignments are then used by \link{createParetoChart}.
}
\examples{
\dontrun{
paretoImportance(pOptions)
}

}
