% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{pImport}
\alias{pImport}
\title{Sample Priority Values}
\format{An object of class \code{numeric} of length 4.}
\usage{
pImport
}
\description{
An example of priority numeric values.
}
\examples{
\dontrun{
 pImport
}
}
\keyword{datasets}
