% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{poImport}
\alias{poImport}
\title{Sample Pareto Option Values}
\format{An object of class \code{numeric} of length 6.}
\usage{
poImport
}
\description{
An example of Pareto option values.
}
\examples{
\dontrun{
 poImport
}
}
\keyword{datasets}
