% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{pOptions}
\alias{pOptions}
\title{Sample Pareto Options}
\format{An object of class \code{character} of length 6.}
\usage{
pOptions
}
\description{
An example of Pareto options.
}
\examples{
\dontrun{
 pOptions
}
}
\keyword{datasets}
