% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{priorities}
\alias{priorities}
\title{Sample Priorities}
\format{An object of class \code{character} of length 4.}
\usage{
priorities
}
\description{
An example of priority inputs.
}
\examples{
\dontrun{
 priorities
}
}
\keyword{datasets}
