% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readParetoOptions}
\alias{readParetoOptions}
\title{Read Pareto Options}
\usage{
readParetoOptions()
}
\value{
This function does not return any values.  Instead, it saves an array character object based
on the user's input to the global environment for use in \link{paretoImportance}.
}
\description{
This is a simple function that provides a prompt to the user to enter in Pareto strategic options. These
will be the x-axis of a Pareto chart in \link{createParetoChart}.
}
\examples{
\dontrun{
readParetoOptions()
}

}
