% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readStrategyOptions}
\alias{readStrategyOptions}
\title{Read Strategy Options}
\usage{
readStrategyOptions()
}
\value{
This function does not return any values.  Instead, it saves an array character object based
on the user's input to the global environment for use in \link{strategyOptionImportance}.  
The object is accessible at \code{sOptions}.
}
\description{
This is a simple function that provides a prompt to the user to enter in strategic options. These
will be the x-axis of a strategy canvas in \link{createStrategyCanvas}.
}
\examples{
readStrategyOptions()
Bookkeeping, Internal Controls Analysis, Financial and Business Review
#Value available in global variable sOptions = c("Bookkeeping","Internal Controls Analysis", Financial and Business Review")

}
