% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readValues}
\alias{readValues}
\title{Read Values}
\usage{
readValues()
}
\value{
This function does not return any values.  Instead, it saves an array character object based
on the user's input to the global environment for use in \link{valuesImportance}.
}
\description{
This is a simple function that provides a prompt to the user to enter in values. These will be 
the names of columns for the priority matrix.  No parameters are taken for this function and all input
values are taken from the console.
}
\examples{
\dontrun{
readValues()
}

}
