% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{soImport1}
\alias{soImport1}
\title{Sample Strategic Options - Industry Importance}
\format{An object of class \code{numeric} of length 6.}
\usage{
soImport1
}
\description{
An example of strategic options importance according to industry.
}
\examples{
\dontrun{
 soImport1
}
}
\keyword{datasets}
