% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{soImport2}
\alias{soImport2}
\title{Sample Strategic Options - Differentiated Importance}
\format{An object of class \code{numeric} of length 6.}
\usage{
soImport2
}
\description{
An example of strategic options importance according to a differentiated competitor.
}
\examples{
\dontrun{
 soImport2
}
}
\keyword{datasets}
