% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{sOptions}
\alias{sOptions}
\title{Sample Strategic Options}
\format{An object of class \code{character} of length 6.}
\usage{
sOptions
}
\description{
An example of strategic options.
}
\examples{
\dontrun{
 sOptions
}
}
\keyword{datasets}
