% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{strategyOptionImportance}
\alias{strategyOptionImportance}
\title{Rank Importance of Strategy Options}
\usage{
strategyOptionImportance(arrayList = c("Option 1", "Option 2", "Option 3"))
}
\arguments{
\item{arrayList}{a list of strategy options to assign numeric importance.  Can be taken from \link{readStrategyOptions} 
output object of \code{sOptions}.}
}
\value{
This function does not return any values.  Instead, it saves an array numeric object based
on the user's input to the global environment.
}
\description{
This function assigns relative rank and importance to the strategy options defined in \link{readStrategyOptions}.  
These numerical assignments are then used by \link{createStrategyCanvas}.
}
\examples{
strategyOptionImportance(sOptions)
}
