% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{values}
\alias{values}
\title{Sample Values}
\format{An object of class \code{character} of length 3.}
\usage{
values
}
\description{
An example of value inputs.
}
\examples{
\dontrun{
 values
}
}
\keyword{datasets}
