% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{vImport}
\alias{vImport}
\title{Sample Value Importance}
\format{An object of class \code{numeric} of length 3.}
\usage{
vImport
}
\description{
An example of values importance.
}
\examples{
\dontrun{
 vImport
}
}
\keyword{datasets}
