% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createHeatmap}
\alias{createHeatmap}
\title{Create Heatmap}
\usage{
createHeatmap(vImport, pImport, priorities, values)
}
\arguments{
\item{vImport}{numeric importance assigned to values from \link{valuesImportance}}

\item{pImport}{numeric importance assigned to priorities from \link{priorityImportance}}

\item{priorities}{a list of priorities from \link{readPriorities}}

\item{values}{a list of values from \link{readValues}}
}
\value{
This function returns a highcharter object (heatmap).
}
\description{
This function creates a heatmap of the priority matrix available in \link{createPriorityMatrix}.
}
\examples{
createHeatmap(vImport,pImport,priorities,values)

}
