% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{optionImportance}
\alias{optionImportance}
\title{Evaluate a Strategic Option Against Priority Matrix}
\usage{
optionImportance(arrayList = c("Professional certifications",
  "On the job learning", "Mentorship"))
}
\arguments{
\item{arrayList}{a list of priorities to evaluate numeric significance for a given strategic option.
Can be taken from \link{readPriorities}.}
}
\value{
This function returns a data frame to evaluate strategic option alignment with a given set of priorities
and values.
}
\description{
This function determines whether a given strategic option is aligned with priorities as defined by the
\link{createPriorityMatrix} and \link{createHeatmap}.
}
\examples{
\dontrun{
optionImportance(priorities)
}

}
