% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{priorityMatrix-package}
\alias{priorityMatrix}
\alias{priorityMatrix-package}
\title{Prioritization Matrix Package}
\description{
The Priortization Matrix package allows users to set strategic priorities using standard tools such as 
a Pareto control chart, Prioritization Matrix (popular in Six Sigma toolkits), and Strategy Canvas (from Blue Ocean Strategy).
}
\details{
The key functions you're likely to need from \pkg{priorityMatrix} are:
\subsection{Prioritzation Matrix}{

\code{\link{readValues}} - a function to input value names

\code{\link{readPriorities}} - a function to input priority names

\code{\link{valuesImportance}} - a function to assign value importance

\code{\link{priorityImportance}} - a function to assign priority importance

\code{\link{createHeatmap}} - a function to create a priority matrix visualization

\code{\link{optionImportance}} - a function to assess fit of a strategic option with priorities and values

}

Otherwise refer to the vignettes.
}
\author{
\strong{Maintainer}: Jordan DeHerrera \email{jordan.deherrera@gmail.com}

}
