% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{valuesImportance}
\alias{valuesImportance}
\title{Rank Importance of Values}
\usage{
valuesImportance(arrayList = c("Contribution", "Certainty", "Connection"))
}
\arguments{
\item{arrayList}{a list of values to assign numeric importance.  Can be taken from \link{readValues} output object of \code{values}.}
}
\value{
This function does not return any values.  Instead, it saves an array numeric object based
on the user's input to the global environment.
}
\description{
This function assigns relative rank and importance to the values defined in \link{readValues}.  These 
numerical assignments are then used by \link{createPriorityMatrix} and \link{createHeatmap}.
}
\examples{
valuesImportance()
# On a scale of 1 to 5, how important is Contribution ?
5
# On a scale of 1 to 5, how important is Certainty ?
2
# On a scale of 1 to 5, how important is Connection ?
5
#Value available in global variable vImport = c(5,2,5)

}
